+++ CameraTools/src/Extensions.cs
@@ -131,6 +131,7 @@ namespace CameraTools
             { "Cycle To Next Cam", "循环到下一个机位" },
             { "Play Current Path", "播放当前路径" },
             { "Move Player With Space Camera", "随太空摄像机移动玩家" },
+            { "Lock Player Position (tmp)", "锁定玩家位置(此项不保存)" },
             { "Path Preview Size", "镜头路径预览大小" },
             { "Reset Windows Position", "重置窗口位置" },